requirements.txt
python-jenkins==1.2.1
sqlalchemy==1.2.11
